+++ subot/settings.py
@@ -45,6 +45,7 @@ class Config:
     project_item: int = 100
     quest: int = 100
     teleportation_shrine: int = 100
+    summoning_brazier: int = 100
 
     detect_objects_through_walls: bool = True
 
@@ -79,6 +80,7 @@ class Config:
             "npc_generic": self.npc_generic,
             "project_item": self.project_item,
             "quest": self.quest,
+            "summoning_brazier": self.summoning_brazier,
             "teleportation_shrine": self.teleportation_shrine,
         }
 
@@ -114,6 +116,7 @@ class Config:
         default_config.project_item = volume.getint("project_item", fallback=default_config.project_item)
         default_config.quest = volume.getint("quest", fallback=default_config.quest)
         default_config.teleportation_shrine = volume.getint("teleportation_shrine", fallback=default_config.teleportation_shrine)
+        default_config.summoning_brazier = volume.getint('summoning_brazier', fallback=default_config.summoning_brazier)
 
         ocr = ini["OCR"]
         default_config.ocr_selected_menu_item = ocr.getboolean("read_selected_menu", fallback=default_config.ocr_selected_menu_item)